+++ src/components/DatePicker/datePickerChangelogData.ts
@@ -10,27 +10,20 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
-    date: "Prerelease",
-    version: "Prerelease",
+    date: "2024-01-04",
+    version: "2.1.4",
     type: "Update",
     affects: ["Styles"],
     notes: ["Chakra 2.8 update."],
   },
+  // TODO: Update on new date
   {
     date: "Prerelease",
     version: "Prerelease",
     type: "Update",
-    affects: ["Styles"],
+    affects: ["Functionality", "Styles"],
     notes: [
       "Fixes bug where month and year calendars were rendering vertically rather than horizontally.",
-    ],
-  },
-  {
-    date: "Prerelease",
-    version: "Prerelease",
-    type: "Update",
-    affects: ["Functionality"],
-    notes: [
       "Updates the `initialDate` and `initialDateTo` props to accept an empty string and adds optional `placeholder` and `placeholderTo` props.",
     ],
   },