+++ src/components/Tabs/Tabs.tsx
@@ -9,12 +9,13 @@ import {
   useMultiStyleConfig,
   ChakraComponent,
 } from "@chakra-ui/react";
-import React, { forwardRef } from "react";
+import React, { forwardRef, useState } from "react";
 
 import Button from "../Button/Button";
 import Icon from "../Icons/Icon";
 import useCarouselStyles from "../../hooks/useCarouselStyles";
 import useNYPLBreakpoints from "../../hooks/useNYPLBreakpoints";
+import useScrollTabIntoView from "../../hooks/useScrollTabIntoView";
 
 // Internal interface used for rendering `Tabs` tab and panel
 // elements, either from data or from children.
@@ -161,6 +162,7 @@ export const Tabs: ChakraComponent<
         useHash = false,
         ...rest
       } = props;
+      const [tabIndex, setTabIndex] = useState(defaultIndex);
       const styles = useMultiStyleConfig("Tabs", {});
       // Just an estimate of the tab width for the mobile carousel.
       const initTabWidth = 65;
@@ -239,13 +241,20 @@ export const Tabs: ChakraComponent<
         );
       }
 
+      const tablistRef = useScrollTabIntoView(tabIndex);
+
       return (
         <ChakraTabs
           defaultIndex={defaultIndex}
           id={id}
           // The following lazy loads each panel whenever it is needed.
           isLazy
-          onChange={onChange}
+          onChange={(index) => {
+            if (onChange !== undefined) {
+              onChange(index);
+            }
+            setTabIndex(index);
+          }}
           ref={ref}
           variant="enclosed"
           {...rest}
@@ -260,7 +269,9 @@ export const Tabs: ChakraComponent<
           >
             {previousButton}
             <Box __css={styles.carouselParent}>
-              <Box {...carouselStyle}>{tabs}</Box>
+              <Box {...carouselStyle} ref={tablistRef}>
+                {tabs}
+              </Box>
             </Box>
             {nextButton}
           </Box>