+++ src/theme/components/feedbackBox.ts
@@ -73,6 +73,7 @@ const FeedbackBox = defineMultiStyleConfig({
       bottom: "0",
       right: "0",
       zIndex: "5",
+      _focus: { boxShadow: "0 0 0 6px var(--nypl-colors-ui-white) !important" },
     },
   }),
 });